Ready to flip to internal stream attachment
